level progress

+++ Scripts/BTS/WebServer/Models/PackageResponces/EventResponce.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 namespace BTS {
     public class EventResponce : PackageResponse {
         public UserModel User { get; private set; }
-        public float Reward { get; private set; }
+        public int Reward { get; private set; }
         public override void Parse(Dictionary<string, object> data) {
             User = new UserModel();
             User.ParseJSON((Dictionary<string, object>)data["user"]);
